feat: add endpoints for suggestions and crud for product materials

+++ backend-quarkus/src/main/java/com/autoflex/inventory/presentation/dto/ProductionSuggestionDTO.java
@@ -1,9 +1,11 @@
 package com.autoflex.inventory.presentation.dto;
 
+import java.math.BigDecimal;
+
 public record ProductionSuggestionDTO(
         String productName,
         Integer quantityToProduce,
-        Double unitPrice,
-        Double totalValue
+        BigDecimal unitPrice,
+        BigDecimal totalValue
 ) {
 }
